+++ src/cli/handlers/publishHandler.ts
@@ -56,6 +56,7 @@ export function createPublishHandler(
 					plugin.app.metadataCache,
 					plugin.settings,
 					plugin.datastore,
+					plugin.extendedCache,
 				);
 
 				const statusManager = new PublishStatusManager(