+++ src/utils/regexes.ts
@@ -6,8 +6,6 @@ export const CODE_FENCE_REGEX = /`(.*?)`/g;
 
 export const CODEBLOCK_REGEX = /```.*?\n[\s\S]+?```/g;
 
-export const EXCALIDRAW_REGEX = /:\[\[(\d*?,\d*?)\],.*?\]\]/g;
-
 export const TRANSCLUDED_SVG_REGEX =
 	/!\[\[(.*?)(\.(svg))\|(.*?)\]\]|!\[\[(.*?)(\.(svg))\]\]/g;